Change Exponential Backoff max_nanos to a whole number.

+++ code/pipeline-cpp/src/config/policy.h
@@ -16,7 +16,7 @@ struct ExponentialBackoff
 
   /// The maximum accumulated wait time in nanoseconds before giving up and
   /// dropping the data.
-  double max_nanos;
+  uint64_t max_nanos;
 
   /// The multiplier to apply to the wait time after each retry. For example, a
   /// multiplier of 2.0 will double the wait time after each retry.
